+++ saros-dlmm/src/amms/test_harness.rs
@@ -9,6 +9,7 @@ use jupiter_amm_interface::{
 };
 use lazy_static::lazy_static;
 
+use saros_sdk::utils::helper::is_swap_for_y;
 use serde_json::{Value, json};
 use solana_account_decoder::{UiAccount, UiAccountEncoding, encode_ui_account};
 use solana_client::{
@@ -32,7 +33,6 @@ use crate::{
     amms::loader::amm_factory,
     route::get_token_mints_permutations,
     swap_instruction::{BuildSwapInstructionDataParams, build_swap_instruction_data},
-    utils::helper::is_swap_for_y,
 };
 use ahash::RandomState;
 use solana_sdk::pubkey;
