+++ saros-sdk/src/state/bin_array.rs
@@ -1,7 +1,7 @@
-use anchor_lang::prelude::ProgramError;
 use anyhow::Result;
 use arrayref::{array_mut_ref, array_ref, array_refs, mut_array_refs};
 use solana_sdk::{
+    program_error::ProgramError,
     program_pack::{IsInitialized, Pack, Sealed},
     pubkey::Pubkey,
 };
