+++ saros-sdk/src/utils/helper.rs
@@ -1,7 +1,6 @@
-use anchor_lang::Key;
 use solana_sdk::pubkey::Pubkey;
 
-use crate::SarosDlmm;
+use crate::constants::{HOOK_CONFIG, HOOK_PROGRAM_ID};
 
 pub fn find_event_authority(program_id: Pubkey) -> Pubkey {
     Pubkey::find_program_address(&[b"__event_authority"], &program_id).0
@@ -15,7 +14,7 @@ pub fn get_bin_array_lower(
     let (bin_array_lower_pubkey, bump) = Pubkey::find_program_address(
         &[
             b"bin_array".as_ref(),
-            pair.key().as_ref(),
+            pair.as_ref(),
             &(bin_array_index).to_le_bytes().as_ref(),
         ],
         program_id,
@@ -31,7 +30,7 @@ pub fn get_bin_array_upper(
     let (bin_array_upper_pubkey, bump) = Pubkey::find_program_address(
         &[
             b"bin_array".as_ref(),
-            pair.key().as_ref(),
+            pair.as_ref(),
             &(bin_array_index + 1).to_le_bytes().as_ref(),
         ],
         program_id,
@@ -42,12 +41,8 @@ pub fn get_bin_array_upper(
 
 pub fn get_hook_bin_array(bin_array_index: u32, pair: &Pubkey) -> (Pubkey, Pubkey) {
     let hook = Pubkey::find_program_address(
-        &[
-            b"hook".as_ref(),
-            SarosDlmm::HOOK_CONFIG.key().as_ref(),
-            pair.key().as_ref(),
-        ],
-        &SarosDlmm::HOOK_PROGRAM_ID.key(),
+        &[b"hook".as_ref(), HOOK_CONFIG.as_ref(), pair.as_ref()],
+        &HOOK_PROGRAM_ID,
     )
     .0;
 
@@ -57,7 +52,7 @@ pub fn get_hook_bin_array(bin_array_index: u32, pair: &Pubkey) -> (Pubkey, Pubke
             hook.as_ref(),
             (bin_array_index).to_le_bytes().as_ref(),
         ],
-        &SarosDlmm::HOOK_PROGRAM_ID.key(),
+        &HOOK_PROGRAM_ID,
     );
     let (hook_bin_array_upper, _) = Pubkey::find_program_address(
         &[
@@ -65,7 +60,7 @@ pub fn get_hook_bin_array(bin_array_index: u32, pair: &Pubkey) -> (Pubkey, Pubke
             hook.as_ref(),
             (bin_array_index + 1).to_le_bytes().as_ref(),
         ],
-        &SarosDlmm::HOOK_PROGRAM_ID.key(),
+        &HOOK_PROGRAM_ID,
     );
 
     (hook_bin_array_lower, hook_bin_array_upper)
